* rc-1.8.0: (39 commits)
  adding transactions to charging stations and pool config and logging improvements
  adding tenantId default to tenant partners
  removing sh files from dockerignore so that entrypoint is available when building
  fixing eslint
  trying to fix license headers
  trying to fix license headers
  trying to fix license headers
  fixing license headers
  .reuse
  Added ignore list 5
  Added ignore list 4
  Added ignore list 3
  Added ignore list 2
  Added ignore list
  License check implementation
  Update protocol when disconnecting
  Introduce ability to force sync DB
  fixed id column for evse table
  Remove unnecessary check for config.util.authProvider.oidc
  OIDC client implementation for callback
  ...

# Conflicts:
#	00_Base/src/config/bootstrap.config.ts
#	00_Base/src/index.ts
#	00_Base/src/interfaces/api/auth/IApiAuthProvider.ts
#	00_Base/src/interfaces/dto/boot.dto.ts
#	00_Base/src/interfaces/dto/certificate.dto.ts
#	00_Base/src/interfaces/dto/change.availability.request.dto.ts
#	00_Base/src/interfaces/dto/charging.profile.dto.ts
#	00_Base/src/interfaces/dto/installed.certificate.dto.ts
#	00_Base/src/interfaces/dto/message.info.dto.ts
#	00_Base/src/interfaces/dto/reservation.dto.ts
#	00_Base/src/interfaces/dto/server.network.profile.dto.ts
#	00_Base/src/interfaces/dto/subscription.dto.ts
#	00_Base/src/interfaces/dto/tariff.dto.ts
#	00_Base/src/interfaces/dto/variable.monitoring.dto.ts
#	00_Base/src/money/Currency.ts
#	00_Base/src/ocpp/model/2.0.1/index.ts
#	00_Base/src/util/MeterValueUtils.ts
#	00_Base/test/money/Currency.test.ts
#	00_Base/test/money/Money.test.ts
#	00_Base/test/util/MeterValueUtils.test.ts
#	01_Data/src/layers/sequelize/model/Certificate/InstalledCertificate.ts
#	01_Data/src/layers/sequelize/model/ChargingStationSecurityInfo.ts
#	01_Data/src/layers/sequelize/model/ChargingStationSequence/ChargingStationSequence.ts
#	01_Data/src/layers/sequelize/model/ChargingStationSequence/index.ts
#	01_Data/src/layers/sequelize/model/Location/ChargingStation.ts
#	01_Data/src/layers/sequelize/model/Location/LatestStatusNotification.ts
#	01_Data/src/layers/sequelize/model/MessageInfo/index.ts
#	01_Data/src/layers/sequelize/model/Tariff/index.ts
#	01_Data/src/layers/sequelize/model/TenantPartner.ts
#	01_Data/src/layers/sequelize/model/VariableMonitoring/EventData.ts
#	01_Data/src/layers/sequelize/model/VariableMonitoring/index.ts
#	01_Data/src/layers/sequelize/repository/ChargingStationSecurityInfo.ts
#	01_Data/src/layers/sequelize/repository/ChargingStationSequence.ts
#	01_Data/src/util/CryptoUtils.ts
#	01_Data/src/util/Pbkdf2.ts
#	01_Data/test/providers/Authorization.ts
#	01_Data/test/providers/Boot.ts
#	01_Data/test/providers/MeterValue.ts
#	02_Util/src/authorization/index.ts
#	02_Util/src/authorizer/index.ts
#	02_Util/src/networkconnection/authenticator/ConnectedStationFilter.ts
#	02_Util/test/providers/ACME.ts
#	02_Util/test/providers/Hubject.ts
#	02_Util/test/util/RequestOperations.test.ts
#	02_Util/test/util/StringOperations.test.ts
#	03_Modules/Configuration/test/providers/SendCall.ts
#	03_Modules/OcppRouter/src/module/router.ts
#	03_Modules/SmartCharging/src/module/smartCharging/InternalSmartCharging.ts
#	03_Modules/Transactions/test/providers/DeviceModelProvider.ts
#	Server/src/config/bootstrap.config.ts
#	Server/src/config/sequelize.bridge.config.js
#	package-lock.json

+++ 01_Data/src/layers/sequelize/util.ts
@@ -1,7 +1,6 @@
-// Copyright (c) 2023 S44, LLC
-// Copyright Contributors to the CitrineOS Project
+// SPDX-FileCopyrightText: 2025 Contributors to the CitrineOS Project
 //
-// SPDX-License-Identifier: Apache 2.0
+// SPDX-License-Identifier: Apache-2.0
 
 import { type BootstrapConfig } from '@citrineos/base';
 import { type Dialect } from 'sequelize';
@@ -108,15 +107,26 @@ export class DefaultSequelizeInstance {
         }
       }
     }
+    this.logger.info(`Sequelize initialized: ${JSON.stringify(this.instance?.config || {})}`);
   }
 
-  private static async syncDb() {
-    if (this.config.database.alter) {
-      await this.instance!.sync({ alter: true });
-      this.logger.info('Database altered');
-    } else if (this.config.database.sync) {
-      await this.instance!.sync({ force: true });
-      this.logger.info('Database synchronized');
+  private static async syncDb(): Promise<void> {
+    if (this.config.database.sync) {
+      const alter = this.config.database.alter;
+      const force = this.config.database.force;
+      if (force) {
+        this.logger.info('Database force synchronizing');
+        await this.instance!.sync({ force: true });
+        this.logger.info('Database force synchronized');
+      } else if (alter) {
+        this.logger.info('Database altering');
+        await this.instance!.sync({ alter: true });
+        this.logger.info('Database altered');
+      } else {
+        this.logger.info('Database synchronizing');
+        await this.instance!.sync();
+        this.logger.info('Database synchronized');
+      }
     }
   }
 
@@ -179,6 +189,7 @@ export class DefaultSequelizeInstance {
         Tenant,
         TenantPartner,
       ],
+      pool: this.config.database.pool,
       logging: (_sql: string, _timing?: number) => {},
     });
   }
